refactor close to work without refresh
