+++ src/index.js
@@ -1,8 +1,5 @@
 import './style.css';
 import './desktop.css';
-import { displayMovies, displayComments } from './modules/fetch.js';
-import { displayAllMovies } from './modules/display.js';
+import { displayMovies } from './modules/fetch.js';
 
 displayMovies();
-displayComments();
-displayAllMovies();
